requirements.txt
# Social Media Analyzer dependencies

# API & HTTP
requests>=2.31.0
httpx>=0.25.0

# Configuration
python-dotenv>=1.0.0

# CLI enhancements (optional)
rich>=13.0.0

# File readers
python-docx>=1.1.0
PyPDF2>=3.0.0

# Web UI
streamlit>=1.29.0
